+++ Changelog.lua
@@ -14,6 +14,7 @@ local changelogMessages = {
     "|cFFFF00Combat Info:|r",
     "[*] Re-enabled GCD Tracker and Potion Timer and updated them to support High Isle.",
     "[*] Fixed small issue with Potion Timer colors.",
+    "[*] Fixed Ultimate display",
     "",
 }